bug 1553314: move and rename processor rules modules
This moves all the Python modules that hold processor rules into the
rules/ directory. Now all the rules are there as well as the base
Rule class.
