+++ socorro/processor/rules/breakpad.py
