+++ socorro/processor/rules/mozilla.py
